requirements.txt
dlib==19.21.0
win10toast==0.9
opencv_python==4.3.0.36
Pillow==8.0.1
